--- controllers/contacts.js
@@ -1,68 +1,0 @@
-// const path = require("path");
-
-const Contact = require("../models");
-const { HttpError, ctrlWrapper } = require("../utils");
-
-// const contactsPath = path.join(__dirname, "../models/contact.js");
-
-const getAll = async (req, res) => {
-  const result = await Contact.find();
-  res.status(200).json(result);
-};
-
-const getById = async (req, res) => {
-  const { contactId } = req.params;
-  // const result = await Contact.findOne({ _id: contactId });
-  const result = await Contact.findById(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json(result);
-};
-
-const add = async (req, res) => {
-  const result = await Contact.create(req.body);
-  res.status(201).json(result);
-};
-
-const updateById = async (req, res) => {
-  const { contactId } = req.params;
-  const { body } = req;
-  const result = await Contact.findByIdAndUpdate(contactId, body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json(result);
-};
-
-const updateStatusContact = async (req, res) => {
-  const { contactId } = req.params;
-  const { body } = req;
-  const result = await Contact.findByIdAndUpdate(contactId, body, {
-    new: true,
-  });
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json(result);
-};
-
-const deleteById = async (req, res) => {
-  const { contactId } = req.params;
-  const result = await Contact.findByIdAndRemove(contactId);
-  if (!result) {
-    throw HttpError(404, "Not found");
-  }
-  res.status(200).json({ message: "contact deleted" });
-};
-
-module.exports = {
-  getAll: ctrlWrapper(getAll),
-  getById: ctrlWrapper(getById),
-  add: ctrlWrapper(add),
-  updateById: ctrlWrapper(updateById),
-  updateStatusContact: ctrlWrapper(updateStatusContact),
-  deleteById: ctrlWrapper(deleteById),
-};
